remove closure from tilemap process
